requirements.txt
# Requirements para Dropi Chile Background Bot
pandas==2.1.4
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Opcional para logs mais avançados
colorlog==6.8.0